Replaced DateTime.Now with DateTime.UtcNow; Changes in the SCTimeOfDayPacket and SCFestivalZoneStatePacket opcodes. In version 2.0 these packages mixed up; Added Transfer Telescop packets; [SQL] fixed auction_house table by Gene; Updated Transfer - fixed carts getting stuck; Some fix.

+++ AAEmu.Game/Core/Packets/G2C/SCTimeOfDayPacket.cs
@@ -5,11 +5,11 @@ namespace AAEmu.Game.Core.Packets.G2C
 {
     public class SCTimeOfDayPacket : GamePacket
     {
-        private readonly ushort _time;
+        private readonly float _time;
 
         public SCTimeOfDayPacket(float time) : base(SCOffsets.SCTimeOfDayPacket, 5)
         {
-            _time = (ushort)time;
+            _time = time;
         }
 
         public override PacketStream Write(PacketStream stream)
